mmc: core: Enable async suspend
Enables the asynchronous suspend of card and class device.  This would
enable both these devices to suspend in separate respective threads, thus
allowing other devices to suspend concurrently.

CRs-fixed: 634333
Change-Id: Ic73022ef777b4fae4b1065f92cfa19da51a99634
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>

mmc: sdhci-msm: Enable async suspend

This is aimed to reduce the suspend time of sdhci-msm device.  The suspend
of this device would be carried out in a separate thread and other devices
may suspend concurrently. This is required since the MMC sub-system
contributes to the major chunk of suspend latency.

CRs-fixed: 634333
Change-Id: I7bdead2deacd96aec5201d82b143b07c470b3892
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>