arm: mm: Support ARCH_MMAP_RND_BITS
arm: arch_mmap_rnd() uses a hard-code value of 8 to generate the
random offset for the mmap base address. This value represents a
compromise between increased ASLR effectiveness and avoiding
address-space fragmentation.  Replace it with a Kconfig option,
which is sensibly bounded, so that platform developers may choose
where to place this compromise. Keep 8 as the minimum acceptable
value.

Bug: 24047224
Change-Id: I2f6c18a0060e1c21b53200ecdcfde9a8c2e3db98
Signed-off-by: Daniel Cashman <dcashman@google.com>