ARM: 8158/1: LLVMLinux: Use static inline in ARM ftrace.h
With compilers which follow the C99 standard (like modern versions of gcc
and clang), "extern inline" does the wrong thing (emits code for an
externally linkable version of the inline function). In this case using
static inline and removing the NULL version of return_address() in
return_address.c does the right thing.

Signed-off-by: Behan Webster <behanw@converseincode.com>
Reviewed-by: Mark Charlebois <charlebm@gmail.com>
Acked-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>

ARM: 8160/1: Drop warning about return_address not using unwind tables

The warning was introduced in 2009 (commit 4bf1fa5 ([ARM] 5613/1:
implement CALLER_ADDRESSx)). The only "problem" here is that
CALLER_ADDRESSx for x > 1 returns NULL which does not do much harm.

The drawback of implementing a fix (i.e. use unwind tables to implement
CALLER_ADDRESSx) is that much of the unwinder code would need to be marked
as not traceable.

Change-Id: I703fddd18ba1529a888c970a548c6e210c2a416b
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>