cpufreq: Share policy from cpu0 to others
This is a workaround to a policy problem on Hammerhead,
where cpu policy sets frequencies incorrectly on additional
cpus.

This patch adds a shared_policy structure, which contains
cpu policy from cpu0. A structure is used to set policies
(min/max frequency and governor) to additional cpus
respectively:

- Define "shared_policy" structure,
- Call __cpufreq_cpu_get with cpu0 argument to gather policy,
- Set min/max frequencies and governor according to
  shared_policy if it is available,
- Free shared_policy after __cpufreq_set_policy work;

This should be used on devices with policy issue only.

Change-Id: Id5e263562dcf1f3868360ea7407b1ee0547c96f7
Signed-off-by: Alex Saiko <solcmdr@gmail.com>