msm: apuclock-krait: Introduce a simple VDD interface
Use the newly added CPUFreq API to create a voltage-on-frequency table
and make it accessible to users.  This can help with a stress balancing
to improve CPU stability, especially on floor frequencies, which can be
used quite often.

Krait power subsystem uses vdd_core variable to set a CPU core voltage.
Now this value can be changed manually from userspace. acpuclk-krait
returns the VDD table in the style of "frequency: voltage".

 ...
 1190mhz: 885 mV
 1267mhz: 895 mV
 1497mhz: 930 mV
 ...

Voltage values are writable and can be changed by passing the whole VDD
table to UV_mV_table sysfs node.  That's not convenient for a terminal
editing, but is evaluated by third-party applications which expect exactly
this implementation.

Both sysfs functions report only the actual frequencies used, leaving the
disabled ones alone.  In addition, write function ensures that the passed
values are within the bounds.  Lower VDD limit comes straight from Krait
Regulator when the upper one is equal to maximum voltage in acpuclk data.

Change-Id: I79c279523268953b72139e9480e00b31ddbf1a36
Signed-off-by: Alex Saiko <solcmdr@gmail.com>

msm: acpuclock-krait: Take voltage boosting into account in VDD table

Currently CPUFreq->VDD table implementation does not know about the
existence of voltage boosting and boost_uv variable, and this can lead
to a couple of issues:

- When boosting is enabled and boost_uv is not zero, it is possible to
  damage regulator bounds and hit a BUG in regulator core code, because
  min_vdd will exceed max_vdd, which is obviously incorrect.

- When boosting is enabled and boost_uv is not zero, real VDDs are not
  equal to the ones printed within VDD table because the values written
  in vdd_core are shown only.

A solution of above problems is simple - we just need to take boost_uv
into account when voltage boosting is actually used. By doing this, the
table will provide the exact information about real voltages on freqs
and operate safely in a proximity of upper VDD bounds.

However, while the first issue is easy to be fixed, the second tampers
the compatibility break up with supported Kernel tweakers. It turns out
that some popular tweakers attemp to mess with the voltage table if the
written value is not the same printed in show_uV_table().  This happens
because boost_uv is simply added to vdd_core value.  That's why it is
not currently possible to solve the second problem without hurting UX.

Unlike maximal VDD limit, a minimal one is not really affected by the
problems states above, so a boosted voltage account there is pointless.

Change-Id: I9387ef521241d8c8068d14171cd28ac490ea17d0
Signed-off-by: Alex Saiko <solcmdr@gmail.com>