net: diag: Support SOCK_DESTROY for inet sockets
This passes the SOCK_DESTROY operation to the underlying protocol diag
handler, or returns -EOPNOTSUPP if that handler does not define a destroy
operation.

Most of this patch is just renaming functions. This is not strictly
necessary, but it would be fairly counterintuitive to have the code to
destroy inet sockets be in a function whose name starts with inet_diag_get.

Change-Id: Iee2c858bf11c48f54890b85b87821a2a2d7109e1
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>