net: diag: Support destroying TCP sockets
This implements SOCK_DESTROY for TCP sockets. It causes all blocking
calls on the socket to fail fast with ECONNABORTED and causes a protocol
close of the socket. It informs the other end of the connection by sending
a RST, i.e., initiating a TCP ABORT as per RFC 793. ECONNABORTED was
chosen for consistency with FreeBSD.

Change-Id: Ice9aad37741fe497341d1d2a51e0b70601a99c90
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>