net: tcp: Deal with listen sockets properly in tcp_abort
When closing a listen socket, tcp_abort currently calls tcp_done without
clearing the request queue.  If the socket has a child socket that is
established but not yet accepted, the child socket is then left without
a parent, causing a leak.

Fix this by setting the socket state to TCP_CLOSE and calling
inet_csk_listen_stop() with the socket lock held, like tcp_close does.

Tested using net_test. With this patch, calling SOCK_DESTROY on a listen
socket that has an established but not yet accepted child socket results
in the parent and the child being closed, such that they no longer appear
in sock_diag dumps.

Change-Id: I2b9c38fab194b3c4e11439047ead9582c811d4c2
Reported-by: Eric Dumazet <eric.dumazet@gmail.com>
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>