ASoC: wcd9xxx: Fix build with !CONFIG_OF
of_property_match_string() call is considred exported, so in case
CONFIG_OF is set to 'false' the build will fail due to a wrapper
will treat 'undefined reference' to that function as a error.

Compile this call out if it is not supported in current kernel build
to avoid this error.  Note that this will make is_enabled() call to
return false anytime for obvious reason.

Change-Id: Ie2359d5752cd253500922756b8f7c2cc66813b97
Signed-off-by: Alex Saiko <solcmdr@gmail.com>